Refactor animation and boxes logging, integrate camera_t and player_t, improve event handling
- animation_t: Add error handling for SDL_IOStream and texture creation; improve logging for loaded frames
- boxes_t: Update logging to include number of frames along with boxes loaded
- build/config: Add player_t and inih to build system and includes
- camera_t: Remove object_t dependency, replace with simple world coordinates; add camera update function linked to player_t
- applicationState_t: Embed camera_t instance; initialize and destroy properly
- main event handling: Refactor event processing; add window resize handling with render scale adjustment; improve event structure and flow

Refactor SDL_AppInit and SDL_AppIterate for improved modularity and initialization

- Split SDL_AppInit into an internal `init` function for cleaner setup logic and better error handling
- Initialize asset loader, settings, SDL subsystems, window, renderer, scaling, vsync, and FPS in a structured order
- Replace applicationState allocation and initialization flow to reduce duplication and improve clarity
- Refactor SDL_AppIterate with an internal `iterate` function; manage static game objects and players initialization only once
- Add detailed loading of objects and players with error handling during iteration
- Integrate vsync begin/end checks with error logs in iteration loop
- Use consistent boolean returns internally, mapping to SDL_AppResult externally
- Fix window creation parameters order and SDL_SetAppMetadata usage

Refactor application quit sequence and add player_t module

- Major refactor of SDL_AppQuit to improve resource cleanup:
  - Introduced `result_t` enum for app exit status abstraction.
  - Created `quit()` helper function to handle quitting FPS, Vsync, settings, asset loader, logging, and SDL resources.
  - Added detailed error logging during quit process.
- Modified object_t rendering to use `camera_t` instead of generic object for camera parameter.
- Added `player_t` module:
  - Defined `player_t` struct wrapping `object_t`.
  - Implemented creation, destruction, state loading, movement, stepping, and rendering functions.
  - Added player_t build config script.
- Minor include and TODO comment additions for future improvements.

Add version and identifier fields to settings\_t; improve memory management and testing for stdfloat16

- Extend `settings_t` structure with `version` and `identifier` string fields for better metadata handling.
- Initialize `version` and `identifier` in `settings_t$create()` with defaults; free them properly in `settings_t$destroy()`.
- Fix macros for 128-bit integer limits in `stdint128.h`.
- Add comprehensive unit tests for `stdfloat16` functions covering bit patterns and conversions.
- Minor includes and TODO comments added for future refactoring.

Add unit tests for 128-bit integer to string conversion functions

- Implement tests for `uint128_to_str` covering zero, small numbers, 64-bit max, and 128-bit max values.
- Implement tests for `int128_to_str` covering zero, small numbers, 64-bit max/min, and 128-bit max/min values.
- Ensure correct string representations for extreme 128-bit integer values.
- Mark tests with TODO for future refactoring.

+++ stdfunc/include/stdint128.h
@@ -5,8 +5,8 @@
 #include "stdfunc.h"
 
 #define UINT128_MAX ( ( uint128_t )( ( uint128_t ) ~( ( uint128_t )0 ) ) )
-#define INT128_MAX ( ( int128_t )( ( int128_t )( ( uint128_t )1 << 127 ) - 1 ) )
-#define INT128_MIN ( ( int128_t )( ( int128_t ) - ( ( uint128_t )1 << 127 ) ) )
+#define INT128_MAX ( ( int128_t )( ( int128_t )( ( int128_t )1 << 127 ) - 1 ) )
+#define INT128_MIN ( ( int128_t )( -( ( int128_t )1 << 127 ) ) )
 
 #define INT128_TYPE_BIT_AMOUNT 128
 
@@ -15,6 +15,7 @@ typedef __int128 int128_t
 typedef unsigned __int128 uint128_t
     __attribute__( ( aligned( BITS_TO_BYTES( INT128_TYPE_BIT_AMOUNT ) ) ) );
 
+// TODO: Refactor all functions
 static FORCE_INLINE char* uint128_to_str( uint128_t value ) {
     // 128-bit max value in base 10 has 39 digits + 1 null terminator
     char* buffer = ( char* )malloc( 40 );
